example/kitchen/appwidget/{,apptheme}: hide image caching behind setter methods
Factor image op caching into setters on the interact structure.

The design for these setters is essentially to bake-in the
`paint.ImageOp` on the first call, as an implementation detail.

Therefore, it's safe to call and re-call every frame, it only allocates
on the first call and every subsequent call is a noop.

I suppose I could have used a pair of `sync.Once`, but that would mean
more fields on the interact.

Signed-off-by: Jack Mordaunt <jackmordaunt.dev@gmail.com>