list,example/unconfigured: add default hooks and example
This commit adds a helper function to generate a list.Hooks
implementation for application developers to use as a
starting point. These hooks will present a message asking
the user to start implementing them, which is kinda fun and
self-referential in a way that I find amusing.

This commit also introduces an example program that displays
this default behavior.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>