list: allow updating Elements in-place
This commit adds logic to the list.processor to permit
providing an updated copy of an element that it is already
managing to the new Update() method. This will replace
the copy that is currently in the Raw list.

Later work will expose this capability to users of
the Manager API.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>