example/kitchen/appwidget/apptheme: eliminate macro
This commit simplifies the room selected indicator by laying the
indicator rectangle on top of the space occupied by the room's
information instead of trying to fit it into a layout next to
the rest of the content. It can be made to work that way, but this
is simple, easy, and readable. We can always pivot if we need the
other behavior.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>