list: prefetch load requests
Prefetch based on a threshold expressed a percentage between [0,1].
Defaults to 15%, clamps at 100%.
Denominator clamped to 1 to hedge against divide by zero from an empty
element list.

NOTE(jfm): if there are no more data to load, it will call tryRequest
quite a lot as it sits in that threshold; is this expected behaviour?

Signed-off-by: Jack Mordaunt <jackmordaunt.dev@gmail.com>