example/kitchen: plumb message data from editor to model
As inline comments suggest, data flow works like this:
1. message is composed in the edit buffer, and submitted via a button
2. on submit, message content is copied from edit buffer and passed to
the backend (RowTracker)
3. backend finalizes the data, and returns it to the caller
4. finalized message data is pushed into the list via list.Update

This process _should_ be async, as the backend system will typically be
an networked api call.

I think now is a good time to start moving the data model to something
realistic.

At the very least, the "backend" api should have an async facade that
mimics a "real" api.

Or perhaps all we need is better naming, but I think you get what I'm
trying to say.

Signed-off-by: Jack Mordaunt <jackmordaunt.dev@gmail.com>