example/kitchen: remove race in Loader implementation
The RowTracker type has a backing slice of list.Element to store
its data. When it's Loader function is invoked, it slices into
that data and returns the relevant subslice. However, this means
that the list.Manager processing those elements is looking at the
same underlying array as the RowTracker, and they race against
one another as they each modify the slice. The list.Manager
moves elements in the slice around, and the RowTracker periodically
sorts it.

This commit introduces a helper that copies the slice before handing
it off to the list.Manager, preventing them from looking at the
same memory.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>