widget,example/kitchen: capture image caching semantics with type
The proper place for this type is debateable.
I think, if the pattern is useful is should be universalised into a type.

The two optional interfaces are to pave the way for opt-in use:

a) images that you don't want to bake every frame, but maybe you want to
be able to change it and have that change get baked
b) images types that implement the image.Image interface can let you
transform them into an image.NRGBA for the sake of performance

The latter point will help when experimenting with `async.Image`, allowing
it to be a little more performant.

Signed-off-by: Jack Mordaunt <jackmordaunt.dev@gmail.com>