list: make manager load during layout, not construction
The manager used to send its first load request in its constructor,
with the rationale being that this would allow the list to already
have populated elements by the time it was first laid out. However,
if you created many managers in parallel, this would result in a lot
of CPU (and I/O if you were loading from somewhere) for data that
wasn't being displayed.

This commit switches to sending that initial load request during
the first frame update, which ensures that managers displaying
state are the only ones performing load operations.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>