example/kitchen: factor MessageStyle out of RowStyle
This commit extracts the logic for laying out a single message
of text atop a background from the logic for laying out a full
row of chat UI. The changes to the appwidget and apptheme
packages are extensive, but surprisingly little needed to
change in the consuming package's state hooks. I think this
demonstrates the flexibility of the current design in a
pleasing way.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>