example/kitchen{,/appwidget/apptheme}: paletted theme
Introduce palette for selecting themes, including a default set of light
and dark themes.

Because of the deficiencies of the material.Theme, some of the color
selection (specifically the luminance checking) needed to be specified
from outside of the material theme package.

In order to get the right colors, most of it was a simple setting mapping
of material.Theme.Fg and material.Theme.Bg.

However, there were some bad assumptions that needed to be fixed up, such
as the scrim color - which needs to be manualy set to a dark color because
it's somewhat theme-agnostic that scrims are dark.

Signed-off-by: Jack Mordaunt <jackmordaunt.dev@gmail.com>