example/kitchen/appwidget/apptheme: mute dark theme text
Having white-on-black text causes some perceptual artifacting and is
generally somewhat unpleasant. This backs the white off to a light
gray. I'd like to go to an even darker gray, but the hotdog ninepatch's
luminance detection uses a light font for some reason, and thus the
contrast of slightly darker grays becomes a problem for readability.

After we fix the luminance detection on ninepatches (assuming that this
is indeed a bug), I suggest we darken this further.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>