list: ensure NoSerial is not mapped to an index
The stateUpdate internal type maps from serials of the state to
their indicies within its array, but was incorrectly mapping the
zero value serial to whatever the last element to have that serial
was.

This commit adds a test enforcing that the zero serial may not
appear within the mapping from serials to indicies.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>