list: ensure removing viewport anchor doesn't jump
Previously, removing the element at list.Position.First would cause
the viewport to jump upward to the top of the already-loaded content.

This happened because the logic tried to find the serial of that
first element within the new state, and it got the zero value as
its index. It would then jump to index zero, which was very
jarring.

This commit introduces a test and a fix for the problem that walks
away from the old beginning of the viewport searching for an
anchor point that it can use to position itself within the new
data. If it can't find any elements that match, it will jump to
the beginning of the new data, but that seems like reasonable
behavior if all of the elements are new.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>