list: only reset list.Position.BeforeEnd on load
This commit changes the position handling so that inserting elements
into the list does not cause the list position to unanchor from the
end if it was sitting at the end before the update. This makes the
manager much more intuitive when building chat-like interfaces in
which you insert new elements and expect to scroll to them. We may
need to add some way to tweak this feature in the future if there
are use-cases where this is undesirable.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>