list: manage insertion carefully at list ends
Previously, it was possible to insert an element into the
underlying list that was very far away from the rest of the
managed content within the underlying data. This broke the
invariant of the list, in that the contents were no longer
contiguous, and this led to subtle bugs.

This commit changes the logic so that insertions to the ends
of the managed list are dropped unless that managed list is
at one of the ends of the underlying data. In that case,
insertions to that end are permitted.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>