widget/plato: let RowStyle branch on message locality
This captures the decision making around message locality in one place.

It makes the API more apparent, previously it was confusing as to the difference
in output between `RowStyle.Local` and `MessageStyle.Local`.

Thus, there is now only `RowStyle.Local` and the decisions are made in the
`RowStyle` constructor.

MessageStyle does a safe default in case it is used directly rather than
through a RowStyle.

Signed-off-by: Jack Mordaunt <jackmordaunt.dev@gmail.com>