widget/plato: style tweaks
Ensure avatar aligns when message is compact, and read-receipt is padded
off the outside the same as the content, but not on in the inside.

This is a balancing act of data vlaues. The data dependencies are
somewhat implicit here because the decisions must coincide but are
made independently.

A better solution is to make the data dependencies more concrete via
structure, but I think this is okay for now.

More precisely, there are two such dependencies:

1. if the avatar is too large or too small it doesn't align vertically
with the bubble
2. the read-receipt must be padded by the same amount as the content,
but there should be no padding between them
	- wrapping them both in the same padding interacts with the macro in
	a way I don't understand

Signed-off-by: Jack Mordaunt <jackmordaunt.dev@gmail.com>