list: provide known sentinel values to indicate start and end of list
Sentinels are implemented as psuedo-elements: independent types that
provide fixed serials.

Note that the serials are not ever sorted because thyey are never
represented as "real" or "managed" elements, unless the synth
implementation decides to re-insert them into the synthetic list.

If that were to be a problem, the Manager could be taught actively
filter-out the setinel values post-synthesis.

A test-harness ensures that the values are sent appropriately.

Signed-off-by: Jack Mordaunt <jackmordaunt.dev@gmail.com>