layout: minimize heap allocation
When number of RowChild is small, use a stack allocated array.

The hope is to define a static array large-enough to avoid needing a
heap allocation for the majority of use-cases.

If however the number of RowChild is large, we fallback to relying on
slice append to allocate as needed.

Signed-off-by: Jack Mordaunt <jackmordaunt.dev@gmail.com>