example/kitchen: enhance data generation api
This is fairly knarly migration to make use of the gen package.

This change was motivated by the desire to establish stronger boundaries
around data generation: to make it more clear where data is being
generated and how.

gen.Generator handles common state, such as maintaining synchronized
serial counters and image fetching callbacks.

The generator methods are indendent because some of the data needs to be
generated asynchronously on-the-fly. Furthermore there's differing
smenatics between generating a "historic"  message that has already been
"sent", and generating a "new" message that has not yet been sent.

Signed-off-by: Jack Mordaunt <jackmordaunt.dev@gmail.com>