example/kitchen: simulate conversations with concurrent actors
To achieve safe concurrency, the `main.Room` gets synchronized.

Each user has an associated goroutine spun up for it that sends messages
to a randomly selected room.

`Send` grows the ability to specify an arbitrary user so that an actor
goroutine can send messages as a specific user.

`model.Room` grows the ability to track whether a user is composing a
message for that room. This will be used to render typing indicators
during synthesis.

Signed-off-by: Jack Mordaunt <jackmordaunt.dev@gmail.com>