async: introduce resource loader
Adapted from Egon's expgio: https://github.com/egonelbre/expgio

Design
===

Loader can queue up blocking operations on a separate goroutine and
avoid blocking the UI.

Blocking operations occur in serial relative to each other, however in
it's current form _some_ of the synchronization may block the UI.

The polling API of `Schedule` allows all interim states to be handled
(queued, loading, loaded), and allows the loader to detect stale data.

If a resource is in-view, it will call Schedule per frame thus telling
the loader to keep it warm. Conversely when a resource is not in-view it
will not call Schedule and will be dallocated as it becomes stale.

Loader is designed to have a useful zero value. Thus, initialization
occurs on the first use.

Synchronization
===

Loader starts a persistent goroutine that processes resources on a queue.
A resource is identified by a tag and has a load function that performs
a potentially long running, blocking operation, such as downloading an
image over the network.

The goroutine processes the queue in a loop, sleeping when there is
nothing to do.

The goroutine is woken up by three events:

1. a new frame is being layed out (provided you have wrapped the UI in a
    call to Frame)
2. a _new_ resource has been scheduled during layout
3. the context has been cancelled (which never happens in the current
    design because context is not exposed to the public API)

Atomics are used when touching a resource because the resources are
accessed inside the processing goroutine _and_ during layout when
scheduled.

Signed-off-by: Jack Mordaunt <jackmordaunt.dev@gmail.com>