widget,example/kitchen{,gen,model}: make kitchen images async
This turned out to be rather complex.

The model now represents images as string URLs.
The layout understands this, and expects to schedule downloads at layout
time.

Because the size of avatars is well known by the layout, we generate a
single 64x64 placeholder on init and use that by default until the real
image is ready.

Since the size of image messages is unknown, we use a heuristic default
size of 320x320 for the placeholder.

In response to feedback on the un-obviousness of the optional Changer
interface used by the CachedImage, I simplified the type and added a
simple Reload method instead.

Reload tells the CachedImage to re-bake on the next call to Cache, since
the underlying image has changed. Once an image has loaded, we Reload
the corresponding CachedImage and pass along the real image.

Layout api can likely be streamlined.

Signed-off-by: Jack Mordaunt <jackmordaunt.dev@gmail.com>