list: prefetch when too few elements
List elements were failing to load when list was initially populated
with very few elements.

When this list has a small number of elements it can fail to enter the
"After" prefetch zone.

This is not a problem in the "Before" direction because index 0 will
always enter the "Before" prefetch zone.

Given 2 elements in a list and a prefetch of 0.15:

- "Before" prefetch zone is 0.15
- "After" prefetch zone is 0.75
- index 1 (second and final element) represents half of the list
    (indexf = 0.5)

Thus, the "After" zone can never be entered since indexf maxes out at
0.5, and "After" requires at least 0.75 of the list to be in-view.

If there are too few elements such that they occupy a fraction greater
than that of the prefetch zones, consider that as having entered the
"After" prefetch zone, and issue a load request accordingly.

A prefetch of 0.15 is roughly 1/7 of the total, therefore any less than
7 elements will incur a load request, regardless of the 0.75 prefetch
zone.

Signed-off-by: Jack Mordaunt <jackmordaunt.dev@gmail.com>