list: expose Stickiness field on Manager
The new list.Manager.Stickiness field gives explicit control
over which list ends (if any) are "sticky" when the scroll
position reaches them. A "sticky" end means that the viewport
will remain attached to that end when new content is inserted
there unless the user scrolls away from that end of the list.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>