widget/material: maintain size for empty images
If the image source is empty, we should maintain the specified size so
that the parent can rely on the Width and Height values they specified.

Prior to this patch, the layout would move around when the image was
empty one frame and present the next. This produces bad user experience
as the layout resizes as images pop in.

Such behaviour reliably occured when using an async.Loader to download
images.

Signed-off-by: Jack Mordaunt <jackmordaunt.dev@gmail.com>