list: [fix] copy state update slice to avoid data race
I observed a data race between layout and the process goroutine that
appears to be a result of sharing the element slice.

While querying for which serials are in the viewport during layout, the
async process was updating those same elements during a modify request.

After reviewing the code I concluded that the slice was being shared and
should rather be copied before reaching layout - that is, before being
sent over the updates channel.

After applying this change, I could not reproduce the race condition.

Signed-off-by: Jack Mordaunt <jackmordaunt.dev@gmail.com>