list: fix idle modification test
The idle modification test was failing because the data
inserted sorted to the end of the list while that direction
was not being ignored. This made the data get discarded.

There were two ways to fix this:

1. Ensure the Comparator hook never indicated that
the data sorted to the end of the list.
2. Trigger load requests to ensure that the async logic
realized that it was at the end of the list.

I chose to implement the first option here, though
the other should also work.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>