list: fix stickiness on frequent state updates
This commit fixes a long-standing and subtle bug in the stickiness
implementation for the list. The way that stickiness works, when the
list.Manager receives a state update, it checks whether the element
at either end of the element list is visible. If it is, it concludes
that it should stick to that end. It calculates whether an element
is visible using the list.Position field of the list passed to
list.Manager.UpdatedLen, and it modifies this position to ensure
that an end is stuck to (if necessary).

The problem with this approach is that the manager can receive
more than one state update during a single frame (it recieves
a slice of them). The first state update could conclude that it
should stick to the end of the list, but the next iteration of the
processing loop would then be looking at the list.Position that
was partially modified by the previous iteration. Its own logic
to check whether an element was visible assumed that the
list.Position was totally valid and consistent, which was no
longer the case since it was altered by the previous loop
iteration. It could conclude that the list should not stick,
and thus the list would come unstuck from the end.

This commit addresses the problem by ensuring that if any of the
update iterations conclude that they should stick, they will
all make that stickiness decision for the frame from that
point forward, regardless of the state of the list.Position.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>